docker/dataset/stock_analysis/requirements-dev.txt
click==7.1.2
cssmin==0.2.0
Flask==1.1.2
Flask-Assets==2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
TA-Lib==0.4.19
webassets==2.0
Werkzeug==1.0.1
